requirements_w2v.txt
matplotlib
datasets
transformers
torchaudio
jiwer
accelerate
PyExecJS
indic-num2words
librosa
soundfile
